refactor: move builders to new module
