+++ pydid/__init__.py
@@ -4,7 +4,8 @@ from .common import DIDError
 from .did import DID, InvalidDIDError
 from .did_url import DIDUrl, InvalidDIDUrlError
 from .doc import DIDDocumentError
-from .doc.doc import DIDDocument, DIDDocumentBuilder, DIDDocumentValidationError
+from .doc.doc import DIDDocument, DIDDocumentValidationError
+from .doc.builder import DIDDocumentBuilder
 from .doc.verification_method import (
     VerificationMethod,
     VerificationSuite,
